+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/BatteryEvent.kt
@@ -8,7 +8,7 @@ internal data class BatteryEvent(
     private val sessionId: String,
     private val batteryPercentage: Float,
     private val isPluggedIn: Boolean,
-    private val metadata: NavigationPerformanceMetadata
+    override var metadata: NavigationPerformanceMetadata
 ) : NavigationPerformanceEvent(sessionId, BATTERY_EVENT_NAME, metadata), Parcelable {
 
     companion object {